example R: add the librpma example
In this example we want to introduce to you a basic way of accessing
Persistent Memory when it is installed in a remote system.
Remote Persistent Memory (or RPMem for short) is a way of doing this
via a network by making use of the Remote Direct Memory Access (RDMA)
technology.

This example is an introduction to RPMem. We will guide you briefly through
setting up all required hardware and software components and verifying whether
the connection works properly. Having that, we will show you how to access
Persistent Memory on a remote system using the librpma library.

After completing this example you will know:
- what is RPMem and what it is good for,
- how RPMem is different comparing to PMem,
- what hardware and software components are required to start using RPMem,
- how to verify whether the RDMA network works properly,
- how to use librpma API to:
- establish a connection,
- prepare memory for remote manipulation,
- manipulate memory on the remote system,
- assure persistency of stores to the remote system.

Step-by-step you will:
- test the connection on the basic level using ping,
- test the connection RDMA capabilities using rping,
- review an application focusing on establishing a connection,
- review an application reading remote system's memory and writing it back
  in the persistent manner.

Co-authored-by: Oksana Salyk <oksana.salyk@intel.com>
Co-authored-by: Jan M Michalski <jan.m.michalski@intel.com>
Co-authored-by: Patryk717 <patrykzbierski717@wp.pl>

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>